requirements/requirements-dev.txt
# Unit tests, coverage, and formatting/style.
pytest==8.0.0
pytest-xdist[psutil]==3.0.2
pytest-cov==5.0.0
ruff==0.3.1
mypy==1.0.0
types-tabulate

# Documentation and examples.
Sphinx==7.2.6
sphinxcontrib-bibtex==2.6.2
sphinx-copybutton==0.5.2
sphinx-autodoc-typehints==2.0.0
myst-nb==1.0.0
myst-parser==3.0.1
pydata-sphinx-theme==0.15.2
jupytext==1.16.1
sphinx-gallery==0.15.0
nbsphinx==0.9.3
matplotlib==3.8.1
pandas==2.1.3
pyscf==2.5.0; sys_platform != 'win32'
openfermion==1.6.1; sys_platform != 'win32'
openfermionpyscf==0.5; sys_platform != 'win32'
bqskit==1.1.1
seaborn==0.13.0
stim==1.13.0
stimcirq==1.13.0
